added database context
