+++ SmartSaver/User.cs
@@ -6,22 +6,11 @@ using System.Data;
 using System.Text;
 using System.Windows.Forms;
 using Microsoft.EntityFrameworkCore;
-using System.Linq;
+
 
 
 namespace SmartSaver
 {
-    public class UserContext : DbContext
-    {
-        public DbSet<User> Users { get; set; }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            optionsBuilder.UseSqlServer("Server=51.75.187.147;Database=SmartSaver;User Id=usern;Password=123456789;");
-        }
-
-    }
-
     public class User
     {
         [Key]
@@ -31,42 +20,6 @@ namespace SmartSaver
         [StringLength(255)]
         public string Password { get; set; }
 
-        public static bool Registration(string email, string pass)
-        {
-            using (var db = new UserContext())
-            {
-                var data = db.Users.Where(a => a.Email == email).ToList();
-
-                if (data.Equals(null))
-                {
-                    db.Add(new User { Email = email, Password = pass });
-                    db.SaveChanges();
-                    return true;
-                }
-                else
-                {
-                    return false;
-                }
-            }
-        }
-        public static bool Login(string email, string pass)
-        {
-            using (var db = new UserContext())
-            {
-                var data = db.Users.Where(a => a.Email == email && a.Password == pass).ToList();
-
-                if (!data.Equals(null))
-                {
-                    return true;
-                }
-                else
-                {
-                    return false;
-                }
-            }
-        }
-
     }
-
 
 }
